requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.3.0
python-dotenv==1.0.0
langchain==0.2.0
langgraph==0.1.0
langchain-openai==0.1.0
python-gitlab==4.6.0
sqlalchemy==2.0.0
asyncpg==0.29.0
pgvector==0.3.0
openai==1.35.0
slack-sdk==3.27.0
apscheduler==3.10.0
httpx==0.27.0
